Support version 1.0.0 of Hashdiff
In version 1.0.0, the `hashdiff` gem changed its module name from
`HashDiff` to `Hashdiff` to prevent name collision with the `hash_diff`
gem (more information:
https://github.com/liufengyun/hashdiff/issues/45).

The API did not change.

This commit updates fixture_builder to allow using the new version of
hashdiff, while maintaining support for the old versions.

It also adds a test for the code path that calls Hashdiff.