Remove date format modification
ActiveSupport has deprecated passing a format identifier to #to_s for
date and time types.  Prior to 7.0.7 calling #to_s with no parameters
would use the `default` format, and the YAML processor calls #to_s on
all fields when dumping content.  FixtureBuilder was setting the
`default` format to the ISO8601 format that the database expects, which
caused the YAML processor to use that content.

As of version 7.0.7 ActiveSupport no longer allows calling #to_s without
parameters on date/time types if a default option is set.  However, the
standard default for Date formats is ISO8601, so FixtureBuilder does not
need to set the default format.  By leaving it unset the YAML dump does
not trigger the deprecation warning, and the resulting formatted strings
are still valid for Postgres.